requirements.txt
starknet-py
fastapi
uvicorn
cairo-lang==0.6.2
pydantic
python-dotenv
